Added MovieManagementController tests

+++ MovieManagement/MovieManagement/Areas/Administration/Controllers/MovieManagementController.cs
@@ -1,14 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.Extensions.Caching.Memory;
 using MovieManagement.Areas.Administration.Models.Movie;
 using MovieManagement.Services.Contracts;
 using MovieManagement.ViewModels;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace MovieManagement.Areas.Administration.Controllers
 {
@@ -32,7 +30,7 @@ namespace MovieManagement.Areas.Administration.Controllers
             var genres = await this.genreService.GetAllGenres();
             model.GenreList = genres.Select(t => new SelectListItem(t.Name, t.Name)).ToList();
 
-            return View(model);
+            return this.View(model);
         }
 
         [HttpPost]
@@ -50,49 +48,72 @@ namespace MovieManagement.Areas.Administration.Controllers
         [HttpGet]
         public IActionResult Delete()
         {
-            return View();
+            return this.View();
         }
 
         [HttpPost]
         public async Task<IActionResult> Delete(string id)
         {
-            var movie = await this.movieService.GetMovieByNameAsync(id);
+            if (this.ModelState.IsValid)
+            {
+                await this.movieService.DeleteMovieAsync(id);
 
-            await this.movieService.DeleteMovieAsync(id);
+                return this.RedirectToAction("TopRated", "Movie");
+            }
 
-            return this.RedirectToAction("TopRated", "Movie");
+            return this.View();
         }
 
         [HttpGet]
         public async Task<IActionResult> Edit(string oldName)
         {
-            var model = await this.movieService.GetMovieByNameAsync(oldName);
-            return View(model);
+            if (this.ModelState.IsValid)
+            {
+                var model = await this.movieService.GetMovieByNameAsync(oldName);
+                return this.View(model);
+            }
+
+            return this.View();
         }
 
         [HttpPost]
         public async Task<IActionResult> Edit(string oldName, MovieViewModel model)
         {
-            var movie = await this.movieService.UpdateMovieAsync(oldName, model);
+            if (this.ModelState.IsValid)
+            {
+                var movie = await this.movieService.UpdateMovieAsync(oldName, model);
 
-            return this.RedirectToAction("Details", "Movie", new { id = movie.Name });
+                return this.RedirectToAction("Details", "Movie", new { id = movie.Name });
+            }
+
+            return this.View(model);
         }
 
         [HttpGet]
         public IActionResult ManageActors(string name)
         {
-            var model = new MovieManageActorsViewModel();
-            model.MovieName = name;
+            if (this.ModelState.IsValid)
+            {
+                var model = new MovieManageActorsViewModel();
+                model.MovieName = name;
 
-            return View(model);
+                return this.View(model);
+            }
+
+            return this.View();
         }
 
         [HttpPost]
         public async Task<IActionResult> ManageActors(MovieManageActorsViewModel model)
         {
-            var movie = await this.movieService.ManageActorAsync(model.MovieName, model.ActorName);
+            if (this.ModelState.IsValid)
+            {
+                var movie = await this.movieService.ManageActorAsync(model.MovieName, model.ActorName);
 
-            return this.RedirectToAction("Details", "Movie", new { id = movie.Name });
+                return this.RedirectToAction("Details", "Movie", new { id = movie.Name });
+            }
+
+            return this.View(model);
         }
     }
 }
